% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{av_glasso}
\alias{av_glasso}
\title{Compute the Adaptive Validation Graphical Lasso}
\usage{
av_glasso(data, seq_r = 0.2 * 0.95^(40:0), C = 0.3)
}
\arguments{
\item{data}{Data matrix of size $n\times d$}

\item{seq_r}{Sequence of considered regularization parameters}

\item{C}{constant C that is used to calibrate the AV regularization parameter}
}
\value{
The selected regularization parameter

The resulting graphical lasso estimate
}
\description{
Compute the Adaptive Validation Graphical Lasso
}
