% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{av_rsme}
\alias{av_rsme}
\title{Computes the adaptive validation regularized score matching estimator rSME (AV)}
\usage{
av_rsme(data, C = 1.5)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{C}{constant C that is used to calibrate the AV regularization parameter}
}
\value{
The selected regularization parameter

The resulting AV regularized score matching estimator rSME (AV)
}
\description{
Computes the adaptive validation regularized score matching estimator rSME (AV)
}
