% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{av_scio}
\alias{av_scio}
\title{Computation of the adaptive validation SCIO}
\usage{
av_scio(data, C = 0.7)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{C}{constant C that is used to calibrate the AV regularization parameter}
}
\value{
The selected regularization parameter

The resulting AV-SCIO estimate
}
\description{
Computation of the adaptive validation SCIO
}
