% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{av_sf_glasso}
\alias{av_sf_glasso}
\title{Computes the graphical lasso with power law regularization and regularization parameter tuned via adaptive validation sf_glasso (AV).}
\usage{
av_sf_glasso(data, C = 1.5)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{C}{Constant C used to tune the thAV. Default value is 1.5}
}
\value{
Graphical lasso with power law regularization and regularization parameter tuned via adaptive validation
}
\description{
Computes the graphical lasso with power law regularization and regularization parameter tuned via adaptive validation sf_glasso (AV).
}
