% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{avtest}
\alias{avtest}
\title{Computes the test that is used in the AV calibration scheme}
\usage{
avtest(estimates, seq_r, counter_r, counter_r_dash, C)
}
\arguments{
\item{estimates}{List of the graphical lasso estimators so far}

\item{seq_r}{List of possible regularization paraeters}

\item{counter_r}{position of the current r in r_seq (see algorithm)}

\item{counter_r_dash}{position of the current r_dash in seq_r (see algorithm)}

\item{C}{constant C that is used for the AV calibration}
}
\value{
TRUE if $l(...) > Cr_dash + Cr$, else FALSE
}
\description{
Computes the test that is used in the AV calibration scheme
}
