% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{bregman_loss}
\alias{bregman_loss}
\title{Computation of the Bregman loss as defined in "Weidong Liu, Xi Luo, Fast and adaptive sparse precision matrix estimation in high dimensions, Journal of Multivariate Analysis, Volume 135, 2015, Pages 153-162"}
\usage{
bregman_loss(sigma, theta)
}
\arguments{
\item{sigma}{Covariance matrix Sigma}

\item{theta}{Precision matrix Theta}
}
\value{
Bregman loss
}
\description{
Computation of the Bregman loss as defined in "Weidong Liu, Xi Luo, Fast and adaptive sparse precision matrix estimation in high dimensions, Journal of Multivariate Analysis, Volume 135, 2015, Pages 153-162"
}
