% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{comparison_similarity}
\alias{comparison_similarity}
\title{Compares the similarity of thAV estimates based on different C.}
\usage{
comparison_similarity(
  n,
  d,
  seq_C = c(0.4, 0.5, 0.6, 0.7, 0.8),
  graph = "random",
  num_reps = 25,
  latex = FALSE
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{seq_C}{Sequence of different C used to be compared}

\item{graph}{Graph type}

\item{num_reps}{Number of repetitions of the recovery task for each recovery method}

\item{latex}{If TRUE, this function returns a latex output via the stargazer package}
}
\value{
Matrix with average (and standard deviation) of F1-score between thAV estimates corresponding to the values in seq_C. In addition, this method also prints the average F1-score between a thAV (based on a specific C) and the true graph. The results are saved in a .txt file.
}
\description{
Compares the similarity of thAV estimates based on different C.
}
