% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{comparisonf1}
\alias{comparisonf1}
\title{Computes a simulation study of the resulting $F_1$-score, Precision and Recall using the oracle, thAV, StARS, and RIC}
\usage{
comparisonf1(
  n,
  d,
  seq_r = 0.2 * 0.95^(40:0),
  C = 0.3,
  graph = NULL,
  lambda = 3,
  num_reps = 50,
  latex = FALSE
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{seq_r}{Set of possible regularization parameters for the thAV}

\item{C}{Constant C used to tune the thAV}

\item{graph}{Type of graph. If graph=NULL, this function does the simulation study for random and scale-free graphs, subsequentially}

\item{lambda}{Threshold factor used to threshold the AV. We thresholdung by lambda*C*r.}

\item{num_reps}{Number of repetitions of the recovery task for each recovery method}

\item{latex}{If TRUE, this function returns a latex output via the stargazer package}
}
\value{
A Matrix that consists of the mean of each performance measure ($F_1$-score, Precision, and Recall) and the standard deviation in brackets
}
\description{
Computes a simulation study of the resulting $F_1$-score, Precision and Recall using the oracle, thAV, StARS, and RIC
}
