% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{f1score}
\alias{f1score}
\title{Compute performance of a precision matrix estimator}
\usage{
f1score(theta, theta.estimate, decimal = 4)
}
\arguments{
\item{theta}{True precision matrix}

\item{theta.estimate}{Estimator of theta}

\item{decimal}{Number of decimals such that entries of absolute size 1e-decimal are set to zero}
}
\value{
f1 F_1-score of theta.estimate with theta being the target

recall Recall of theta.estimate with theta being the target

precision Precision of theta.estimate with theta being the target
}
\description{
Compute performance of a precision matrix estimator
}
