% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_precmatrix.R
\name{generateGraph}
\alias{generateGraph}
\title{Generates a precision matrix list consisting of a precision matrix and a igraph layout}
\usage{
generateGraph(d, graph = "random", rangemin = 0.5, rangemax = 0.9, prob = NULL)
}
\arguments{
\item{d}{Dimension of the precision matrix}

\item{graph}{Graph type}

\item{rangemin, }{rangemax The generation of the entries is based on sampling from a uniform distribution from [-rangemax, -rangemin] and [-rangemin, rangemax]}

\item{prob}{Probability that is used to generate the Gilbert graph, which defines the adjacency matrix of the graph}
}
\value{
list containing the precision matrix and an igraph layout
}
\description{
Generates a precision matrix list consisting of a precision matrix and a igraph layout
}
\examples{
rd <- generateGraph(100)
network_rd <- graph.adjacency(rd[[1]], mode="undirected", weighted=TRUE, diag=FALSE)
plot.igraph(network_rd, layout=rd[[2]])

sf <- generateGraph(100, graph="scale-free")
network_sf <- graph.adjacency(sf[[1]], mode="undirected", weighted=TRUE, diag=FALSE)
plot.igraph(network_sf, layout=sf[[2]])
}
