% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{graph_comparison}
\alias{graph_comparison}
\title{Compares the resulting graphs for several estimation methods (true graph, oracle, StARS, scaled lasso, TIGER, rSME (eBIC), SCIO (CV), SCIO (bregman), and thAV)}
\usage{
graph_comparison(n, d, type = "random")
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{type}{Graph type}
}
\value{
Saves plots of the graphs and returns F1-scores
}
\description{
Compares the resulting graphs for several estimation methods (true graph, oracle, StARS, scaled lasso, TIGER, rSME (eBIC), SCIO (CV), SCIO (bregman), and thAV)
}
