% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{infloss}
\alias{infloss}
\title{Compute the infloss on the off-diagonals of two matrices}
\usage{
infloss(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Matrix with the same size as A}
}
\value{
Maximal off-diagonal difference between A and B
}
\description{
Compute the infloss on the off-diagonals of two matrices
}
