% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{normalize_theta}
\alias{normalize_theta}
\title{Turns a covariance matrix into the correlation matrix}
\usage{
normalize_theta(theta)
}
\arguments{
\item{theta}{Covariance matrix}
}
\value{
Correlation matrix
}
\description{
Turns a covariance matrix into the correlation matrix
}
