% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{penalty_matrix}
\alias{penalty_matrix}
\title{Computes the reweighted l1 regularization to compute the graphical lasso with power law regularization (Liu & Ihler(2011)).}
\usage{
penalty_matrix(est, alpha)
}
\arguments{
\item{est}{Estimation of Theta from the previous iteration}

\item{alpha}{Regularization parameter}
}
\value{
The reweighted penalties as a dxd matrix.
}
\description{
Computes the reweighted l1 regularization to compute the graphical lasso with power law regularization (Liu & Ihler(2011)).
}
