% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{performance_thav}
\alias{performance_thav}
\title{Computes the performance of the thAV for large-scale examples}
\usage{
performance_thav(
  d,
  n = 500,
  seq_r = seq(0.05, 0.4, length.out = 40),
  C = 0.7,
  graph = "random",
  lambda = 1,
  num_reps = 20,
  latex = TRUE
)
}
\arguments{
\item{d}{Number of dimensions}

\item{n}{Number of samples}

\item{seq_r}{Set of possible regularization parameters for the thAV}

\item{C}{Constant C used to tune the thAV}

\item{graph}{Type of graph. If graph=NULL, this function does the simulation study for random and scale-free graphs, subsequentially}

\item{lambda}{Threshold factor used to threshold the AV. We thresholdung by lambda*C*r.}

\item{num_reps}{Number of repetitions of the recovery task for each recovery method}

\item{latex}{If TRUE, this function returns a latex output via the stargazer package}
}
\value{
A Matrix that consists of the mean of each performance measure (F1-score, Precision, and Recall) and the standard deviation in brackets.
}
\description{
Computes the performance of the thAV for large-scale examples
}
