% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{precision}
\alias{precision}
\title{Compute precision of a precision matrix estimator}
\usage{
precision(theta, theta.estimate)
}
\arguments{
\item{theta}{True precision matrix}

\item{theta.estimate}{Estimator of theta}
}
\value{
Precision of theta.estimate with theta being the target
}
\description{
Compute precision of a precision matrix estimator
}
