% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{recall}
\alias{recall}
\title{Compute recall of a precision matrix estimator}
\usage{
recall(theta, theta.estimate)
}
\arguments{
\item{theta}{True precision matrix}

\item{theta.estimate}{Estimator of theta}
}
\value{
Recall of theta.estimate with theta being the target
}
\description{
Compute recall of a precision matrix estimator
}
