% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{recovery_topvalues_split}
\alias{recovery_topvalues_split}
\title{Simulation study about the recovery of the most significant precision matrix entries
We split the simulation into several parts due to memory issues.}
\usage{
recovery_topvalues_split(
  n,
  d,
  part,
  perc_top = 0.2,
  num_reps = 10,
  graph = "random"
)
}
\arguments{
\item{n}{Sample size}

\item{d}{Size of the graph}

\item{part}{Number that describes the part of the simulation study}

\item{perc_top}{Percentage of most significant precision matrix entries that are taken into consideration}

\item{num_reps}{Number of repetitions which are run during the simulation}

\item{graph}{Type of the graph}
}
\value{
Writes a file that is read by the recovery_topvalues function to summarize the experiment
}
\description{
Simulation study about the recovery of the most significant precision matrix entries
We split the simulation into several parts due to memory issues.
}
