% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{scio_bregman}
\alias{scio_bregman}
\title{Computation of the Sparse Columnwise Inverse Operator (SCIO) estimator tuned via the Bregman loss: We pick the regularization parameter that minimizes the Bregman loss.}
\usage{
scio_bregman(data)
}
\arguments{
\item{data}{Data matrix of size n x d}
}
\value{
SCIO estimator
}
\description{
Computation of the Sparse Columnwise Inverse Operator (SCIO) estimator tuned via the Bregman loss: We pick the regularization parameter that minimizes the Bregman loss.
}
