% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{score_ebic}
\alias{score_ebic}
\title{Computes the rSME via eBIC}
\usage{
score_ebic(data, gamma = 1)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{gamma}{Parameter that is used in the eBIC criterion. Usually gamma=0.5 or gamma=1. The default value is set to 1.}
}
\value{
rSME estimator that is tuned via eBIC
}
\description{
Computes the rSME via eBIC
}
