% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{sf_glasso}
\alias{sf_glasso}
\title{Computes the graphical lasso with power law regularization sf_glasso (Liu & Ihler(2011)).}
\usage{
sf_glasso(data, alpha)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{alpha}{Regularization parameter}
}
\value{
Graphical lasso with power law regularization
}
\description{
Computes the graphical lasso with power law regularization sf_glasso (Liu & Ihler(2011)).
}
