% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{simu_av_f1}
\alias{simu_av_f1}
\title{Runs a simulation study of the resulting F1-score, Precision and Recall using only thAV.}
\usage{
simu_av_f1(
  n_seq,
  d_seq,
  estimator = "thav_glasso",
  graph = "random",
  num_reps = 10,
  latex = TRUE
)
}
\arguments{
\item{n_seq}{Sequence of number of samples ((n_seq[j], d_seq[j]) corresponds to one setting)}

\item{d_seq}{Sequence of number of dimensions ((n_seq[j], d_seq[j]) corresponds to one setting)}

\item{estimator}{Chosen estimator. Must be either "thav_glasso", "thav_rsme", or "av_scio"}

\item{graph}{Type of graph}

\item{num_reps}{Number of repetitions of the recovery task for each recovery method}

\item{latex}{If TRUE, this function returns a latex output using the stargazer package}
}
\value{
A tex-ready code for a table that consists of the mean of each performance measure (F1-score, Precision, and Recall) and the standard deviation in brackets. The result is saved as a .txt file.
}
\description{
Runs a simulation study of the resulting F1-score, Precision and Recall using only thAV.
}
