% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{simu_f1}
\alias{simu_f1}
\title{Reads the .txt files created by the simu_f1_split function to summarize the resulting F1-score, Precision and Recall of the oracle, StARS, scaled Lasso, TIGER, rSME using eBIC with gamma=1, rSME using eBIC with gamma=0.5, SCIO using Cross-Validation, SCIO using the bregman loss, thAV-glasso, AV-rSME, and AV-SCIO. It is necessary to run simu_f1_spit beforehand.}
\usage{
simu_f1(n, d, graph = "random", num_parts = 3)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{graph}{Type of graph, "random" or "scale-free"}

\item{num_parts}{Number of parts in which we split the simulation study}
}
\value{
A tex-ready code for a table that consists of the mean of each performance measure (F1-score, Precision, and Recall) and the standard deviation in brackets. The result is saved as a .txt file.
}
\description{
Reads the .txt files created by the simu_f1_split function to summarize the resulting F1-score, Precision and Recall of the oracle, StARS, scaled Lasso, TIGER, rSME using eBIC with gamma=1, rSME using eBIC with gamma=0.5, SCIO using Cross-Validation, SCIO using the bregman loss, thAV-glasso, AV-rSME, and AV-SCIO. It is necessary to run simu_f1_spit beforehand.
}
