% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{simu_f1_split}
\alias{simu_f1_split}
\title{Computes parts of the simulation study that compares F1-score, Precision and Recall using the oracle, StARS, scaled Lasso, TIGER, rSME using eBIC with gamma=1, rSME using eBIC with gamma=0.5, SCIO using Cross-Validation, SCIO using the bregman loss, thAV-glasso, AV-rSME, and AV-SCIO.
We split the simulation into several parts because of memory issues.}
\usage{
simu_f1_split(n, d, part, num_reps = 10, graph = "random", C_thav = 0.7)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{part}{Number that describes the part of the simulation study}

\item{num_reps}{Number of repetitions (in this part) of the recovery task for each recovery method}

\item{graph}{Type of graph.}

\item{C_thav}{Constant C used to tune the thAV}

\item{C_score}{Constant C used to tune the AV-rSME}

\item{C_scio}{Constant C used to tune the AV-SCIO}
}
\value{
Writes a .txt-file that can be read by the simu_f1 function to summarize the simulation study.
}
\description{
Computes parts of the simulation study that compares F1-score, Precision and Recall using the oracle, StARS, scaled Lasso, TIGER, rSME using eBIC with gamma=1, rSME using eBIC with gamma=0.5, SCIO using Cross-Validation, SCIO using the bregman loss, thAV-glasso, AV-rSME, and AV-SCIO.
We split the simulation into several parts because of memory issues.
}
