% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{simu_thav_others}
\alias{simu_thav_others}
\title{Computes F1-score, Precision and Recall using the thAV-version of the rSME estimator or the glasso with power law regularization (proposed by Liu & Ihler (2011). 
The results are averaged over several repetitions.}
\usage{
simu_thav_others(
  n,
  d,
  estimator,
  C = 1.5,
  lambda = 1.5,
  num_reps = 10,
  graph = "scale-free"
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{estimator}{Type of estiamtor. Either thav_rsme or thav_sf}

\item{C}{Constant C used to tune the thAV. Default value is 1.5}

\item{lambda}{Thresholding parameter lambda. Default value is 1.5}

\item{num_reps}{Number of repetitions of the recovery task. Default value is 10.}

\item{graph}{Type of graph. One can choose between "scale-free" and "random". Default value is "scale-free".}
}
\value{
Writes a .txt-file that can be read by the comparisonf1 function to summarize the simulation study.
}
\description{
Computes F1-score, Precision and Recall using the thAV-version of the rSME estimator or the glasso with power law regularization (proposed by Liu & Ihler (2011). 
The results are averaged over several repetitions.
}
