% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{std}
\alias{std}
\title{Returns the standard deviation of a vector}
\usage{
std(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Standard deviation of vector x
}
\description{
Returns the standard deviation of a vector
}
