% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{thAV.estimator}
\alias{thAV.estimator}
\title{Compute the thAV estimator}
\usage{
thAV.estimator(data, seq_r = 0.2 * 0.95^(40:0), C = 0.3, lambda = 3)
}
\arguments{
\item{data}{Data matrix of size $n\times d$}

\item{seq_r}{Sequence of considered regularization parameters}

\item{C}{constant C that is used to calibrate the AV regularization parameter}

\item{lambda}{Threshold factor lambda: We threshold the AV solution by lambda*C*av$TuningParameter}
}
\value{
The thresholded AV estimator
}
\description{
Compute the thAV estimator
}
