% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{unregularized_threshold}
\alias{unregularized_threshold}
\title{Computes the performance (F1, Precision, and Recall) of the thresholded Maximum-Likelihood estimator using different thresholds. The generated .txt-files are used for the adaptation_threshold.ipynb to create a plot.}
\usage{
unregularized_threshold(n, d, graph = "random")
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{graph}{Type of graph}
}
\value{
Saves .txt-files that will be used by adaptation_threshold.ipynb to generate a plot
}
\description{
Computes the performance (F1, Precision, and Recall) of the thresholded Maximum-Likelihood estimator using different thresholds. The generated .txt-files are used for the adaptation_threshold.ipynb to create a plot.
}
