% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{visualization_differentC}
\alias{visualization_differentC}
\title{Compares the resulting graphs of thAV with different C}
\usage{
visualization_differentC(
  n,
  d,
  seq_C = c(0.5, 0.6, 0.7, 0.8),
  graph = "random",
  add_title = FALSE
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{seq_C}{Sequence of different C used to be compared}

\item{graph}{Graph type}

\item{add_titlw}{If TRUE, all plots have a title}
}
\value{
Plots of the different thAV graphs that are computet using C from seq_C. In addition, this method prints a table with the F1 score between the truth and each thAV estimate.
}
\description{
Compares the resulting graphs of thAV with different C
}
