% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{weight_dist}
\alias{weight_dist}
\title{Investigate the empirical distribution of the absolute values of the graphs edges}
\usage{
weight_dist(d, graph = "random", num_reps = 50)
}
\arguments{
\item{d}{Number of dimensions}

\item{graph}{Graph type}

\item{num_reps}{Number of repetitions}
}
\value{
Saves a .csv-file that can be read by visualizations_weights.ipynb to plot a histogram of the weight distribution
}
\description{
Investigate the empirical distribution of the absolute values of the graphs edges
}
