% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{adaptation_threshold}
\alias{adaptation_threshold}
\title{Computes the performance (F_1, Precision, and Recall) of thAV using several thresholds, and of StARS and RIC.}
\usage{
adaptation_threshold(n, d, C = 0.3, graph = "random")
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{C}{Constant C used to tune the thAV}

\item{graph}{Type of graph. If graph=NULL, this function does the simulation study for random and scale-free graphs, subsequentially}
}
\value{
Saves .csv-files that will be used by adaptation_threshold.ipynb to generate a plot
}
\description{
Computes the performance (F_1, Precision, and Recall) of thAV using several thresholds, and of StARS and RIC.
}
