% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_estimation.r
\name{av_glasso}
\alias{av_glasso}
\title{Computes the adaptive validation graphical lasso}
\usage{
av_glasso(data, C = 0.7)
}
\arguments{
\item{data}{Data matrix of size n x d}

\item{C}{constant C that is used to calibrate the AV regularization parameter}
}
\value{
The selected regularization parameter

The resulting graphical lasso estimate
}
\description{
Computes the adaptive validation graphical lasso
}
\examples{
theta <- generateGraph(d=100)[[1]]
Sigma <- solve(theta)
data <- scale(mvrnorm(n=200, mu=rep(0, 100), Sigma=Sigma))
av <- av_glasso(data)
av_tp <- av$TuningParameter
av_est <- av$Thetahat
}
