% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{comparison_density}
\alias{comparison_density}
\title{Reads the .txt files created by the comparison_density_split function to summarize the resulting F1-score, Precision and Recall using the oracle, thAV, StARS, scaled lasso, and TIGER.}
\usage{
comparison_density(n, d, seq_density, graph, num_parts)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{seq_density}{Sequence with values for p, which is the probability to connect 2 edges of a Gilbert graph (this is the way how the random graph is generated).}

\item{graph}{Type of graph. If graph=NULL, this function does the simulation study for random and scale-free graphs, subsequentially}

\item{num_parts}{Number of parts in which we split the simulation study}
}
\value{
Matrix that measures average (and standard deviation) of the F1-score between the true graph and an estimation (thAV, StARS, and RIC)
}
\description{
Reads the .txt files created by the comparison_density_split function to summarize the resulting F1-score, Precision and Recall using the oracle, thAV, StARS, scaled lasso, and TIGER.
}
