% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{graph_comparison}
\alias{graph_comparison}
\title{Compares the resulting graphs for several estimation methods}
\usage{
graph_comparison(n, d, type = "random")
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{type}{Graph type}
}
\value{
Saves plots and $F_1$-scores of the true graph, the oracle, the thAV, the thAV2 (lambda=2), the StARS, and the RIC
}
\description{
Compares the resulting graphs for several estimation methods
}
