% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{performance_thav}
\alias{performance_thav}
\title{Compute the performance of the thAV for large-scale examples}
\usage{
performance_thav(
  d,
  n = 500,
  seq_r = 0.2 * 0.95^(40:0),
  C = 0.3,
  graph = "random",
  lambda = 3,
  num_reps = 20,
  latex = TRUE
)
}
\arguments{
\item{d}{Number of dimensions}

\item{n}{Number of samples}

\item{seq_r}{Set of possible regularization parameters for the thAV}

\item{C}{Constant C used to tune the thAV}

\item{graph}{Type of graph. If graph=NULL, this function does the simulation study for random and scale-free graphs, subsequentially}

\item{lambda}{Threshold factor used to threshold the AV. We thresholdung by lambda*C*r.}

\item{num_reps}{Number of repetitions of the recovery task for each recovery method}

\item{latex}{If TRUE, this function returns a latex output via the stargazer package}
}
\value{
A Matrix that consists of the mean of each performance measure ($F_1$-score, Precision, and Recall) and the standard deviation in brackets.
}
\description{
Compute the performance of the thAV for large-scale examples
}
