% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{recovery_topvalues}
\alias{recovery_topvalues}
\title{Simulation study about the recovery of the most significant precision matrix entries.
This function concatinates the results of recovery_topvalues_split and returns a summary.}
\usage{
recovery_topvalues(n_seq, d_seq, num_parts = c(2, 3, 2), graph = "random")
}
\arguments{
\item{n_seq}{Sequence of sample size}

\item{d_seq}{Sequence of dimensions}

\item{num_parts}{Sequence of number of parts, in each setting n_seq[j], d_seq[j]}

\item{graph}{Type of the graph}
}
\value{
Mean and standard deviation of the proportional amount of most significant edges that got recovered
}
\description{
Simulation study about the recovery of the most significant precision matrix entries.
This function concatinates the results of recovery_topvalues_split and returns a summary.
}
