% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{thav_topvalues}
\alias{thav_topvalues}
\title{Exemplary recovery in F_1-score of the top edges for several settings}
\usage{
thav_topvalues(n_seq, d_seq, num_top = 50, graph = "random", latex = TRUE)
}
\arguments{
\item{n_seq}{Sequence of sample sizes used for the simulation}

\item{d_seq}{Sequence of dimensionaly of the precision matrix used for the simulation}

\item{num_top}{Amount of entries that are going to be analyzed}

\item{graph}{Type of the graph}

\item{latex}{If TRUE, returns a latex output}
}
\value{
Table including the largest, the num_top'th largest absolute value of Theta, the reference value 6Cr, and the proportional amount of recovered top edges for several settings
}
\description{
Exemplary recovery in F_1-score of the top edges for several settings
}
