% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_simulations.R
\name{visualization_differentC}
\alias{visualization_differentC}
\title{Compare the resulting graphs of thAV with different C}
\usage{
visualization_differentC(
  n,
  d,
  seq_C = c(0.2, 0.3, 0.4, 0.5),
  graph = "random",
  add_title = TRUE
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of dimensions}

\item{seq_C}{Sequence of different C used to be compared}

\item{graph}{Graph type}

\item{add_titel}{If TRUE, all plots have a title}
}
\value{
Plots of the different thAV graphs that are computet using C from seq_C. In addition, this method prints a table with the f_1 score between the truth and each thAV estimate.
}
\description{
Compare the resulting graphs of thAV with different C
}
